x/bank/simulation/decoder_test.go
package simulation_test

import (
	"testing"

	"github.com/stretchr/testify/require"

	"github.com/cosmos/cosmos-sdk/simapp"

	"github.com/cosmos/cosmos-sdk/types/kv"
	"github.com/cosmos/cosmos-sdk/x/bank/simulation"
)

func TestDecodeStore(t *testing.T) {
	app := simapp.Setup(false)
	dec := simulation.NewDecodeStore(app.BankKeeper)

	//totalSupply := types.NewSupply(sdk.NewInt64Coin(sdk.DefaultBondDenom, 1000))
	//
	//supplyBz, err := app.BankKeeper.MarshalSupply(totalSupply)
	//require.NoError(t, err)

	kvPairs := kv.Pairs{
		Pairs: []kv.Pair{
			//{Key: types.SupplysPrefix + []byte(sdk.DefaultBondDenom), Value: supplyBz},
			{Key: []byte{0x99}, Value: []byte{0x99}},
		},
	}

	tests := []struct {
		name        string
		expectedLog string
	}{
		//{"Supply", fmt.Sprintf("%v\n%v", totalSupply, totalSupply)},
		{"other", ""},
	}

	for i, tt := range tests {
		i, tt := i, tt
		t.Run(tt.name, func(t *testing.T) {
			switch i {
			case len(tests) - 1:
				require.Panics(t, func() { dec(kvPairs.Pairs[i], kvPairs.Pairs[i]) }, tt.name)
			default:
				require.Equal(t, tt.expectedLog, dec(kvPairs.Pairs[i], kvPairs.Pairs[i]), tt.name)
			}
		})
	}
}
